requirements.txt
asgiref==3.8.1
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
joblib==1.4.2
numpy==2.1.1
py4j==0.10.9.7
scikit-learn==1.5.1
scipy==1.14.1
sqlparse==0.5.1
threadpoolctl==3.5.0
tzdata==2024.1
